requirements.txt
asgiref==3.9.2
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Babel==2.10.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
distlib==0.4.0
Django==5.2.6
django-cors-headers==4.9.0
django-crispy-forms==2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
filelock==3.19.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
MarkupSafe==3.0.2
msgpack==1.1.1
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
platformdirs==4.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
virtualenvwrapper-win==1.2.7
Werkzeug==3.1.3
zope.interface==8.0.1
